#18: Restructuring pkg code (docker, mvn, config, etc)
