+++ cmd/build.go
@@ -3,8 +3,8 @@ package cmd
 import (
 	"os"
 
-	"github.com/ahstn/atlas/builder"
 	"github.com/ahstn/atlas/flag"
+	"github.com/ahstn/atlas/pkg/builder"
 	"github.com/apex/log"
 	"github.com/urfave/cli"
 )
