+++ cmd/project.go
@@ -3,9 +3,9 @@ package cmd
 import (
 	"path"
 
-	"github.com/ahstn/atlas/builder"
-	"github.com/ahstn/atlas/config"
 	"github.com/ahstn/atlas/flag"
+	"github.com/ahstn/atlas/pkg/builder"
+	"github.com/ahstn/atlas/pkg/config"
 	"github.com/urfave/cli"
 	emoji "gopkg.in/kyokomi/emoji.v1"
 )
