+++ flag/clean.go
@@ -3,6 +3,6 @@ package flag
 import "github.com/urfave/cli"
 
 var Clean = cli.BoolFlag{
-	Name:  "c, clean",
+	Name:  "clean, c",
 	Usage: "clean artifacts before building",
 }
