+++ flag/config.go
@@ -5,7 +5,7 @@ import "github.com/urfave/cli"
 // Config handles passing a file's path
 // TODO: Validation, i.e. file ends in .yaml
 var Config = cli.StringFlag{
-	Name:  "c, config",
+	Name:  "config, c",
 	Value: "atlas.yaml",
 	Usage: "name of config file in ~/.config/atlas",
 }
