+++ flag/verbose.go
@@ -3,6 +3,6 @@ package flag
 import "github.com/urfave/cli"
 
 var Verbose = cli.BoolFlag{
-	Name:  "V, verbose",
+	Name:  "verbose, V",
 	Usage: "verbose logging rather than progress bars",
 }
